api/requirements.txt
flask==2.0.1
werkzeug==2.0.1
google-generativeai==0.3.1
python-dotenv==0.19.0
requests==2.27.1
Pillow==10.1.0
python-multipart==0.0.5 